State

Types

Stopping.GenericState — Type

Type: GenericState

Methods: update!, reinit!

A generic State to describe the state of a problem at a point x.

Tracked data include:

x : current iterate

d [opt] : search direction

res [opt] : residual

current_time [opt] : time

current_score [opt] : score

Constructor: GenericState(:: AbstractVector; d :: Iterate = nothing, res :: Iterate = nothing, current_time :: FloatVoid = nothing, current_score :: Iterate = nothing)

Note: By default, unknown entries are set to nothing.

Examples: GenericState(x) GenericState(x, current_time = 1.0) GenericState(x, current_score = 1.0)

See also: Stopping, NLPAtX

Stopping.ListStates — Type

Type: list of States

Constructor:

ListStates(:: AbstractState)

Note:

If n != -1, then it stores at most n AbstractState.

add additional methods following https://docs.julialang.org/en/v1/base/collections/

ListStates recursively handles sub-list of states as the attribute list is

an array of pair whose first component is a, AbstractState and the second component is a ListStates (or nothing).

Examples: ListStates(state) ListStates(state, n = 2) ListStates(n = -1, list = [[state1, nothing], [state2, nothing]], i = 2) ListStates(n = -1, list = [[state1, another_list]], i = 1)

Stopping.NLPAtX — Type

Type: NLPAtX

Methods: update!, reinit!

NLPAtX contains the information concerning a nonlinear optimization model at the iterate x.

min_{x ∈ ℜⁿ} f(x) subject to lcon <= c(x) <= ucon, lvar <= x <= uvar.

Tracked data include:

x : the current iterate

fx [opt] : function evaluation at x

gx [opt] : gradient evaluation at x

Hx [opt] : hessian evaluation at x

mu [opt] : Lagrange multiplier of the bounds constraints

cx [opt] : evaluation of the constraint function at x

Jx [opt] : jacobian matrix of the constraint function at x

lambda : Lagrange multiplier of the constraints

d [opt] : search direction

res [opt] : residual

current_time [opt] : time

current_score [opt] : score

evals [opt] : number of evaluations of the function (import the type NLPModels.Counters)

Constructor: NLPAtX(:: AbstractVector, :: AbstractVector; fx :: FloatVoid = nothing, gx :: Iterate = nothing, Hx :: MatrixType = nothing, mu :: Iterate = nothing, cx :: Iterate = nothing, Jx :: MatrixType = nothing, current_time :: FloatVoid = nothing, current_score :: Iterate = nothing, evals :: Counters = Counters())

NLPAtX(:: AbstractVector; fx :: FloatVoid = nothing, gx :: Iterate = nothing, Hx :: MatrixType = nothing, mu :: Iterate = nothing, current_time :: FloatVoid = nothing, current_score :: Iterate = nothing, evals :: Counters = Counters())

Note: - By default, unknown entries are set to nothing (except evals). - All these information (except for x and lambda) are optionnal and need to be update when required. The update is done through the update! function. - x and lambda are mandatory entries. If no constraints lambda = []. - The constructor check the size of the entries.

See also: GenericState, update!, update_and_start!, update_and_stop!, reinit!

Stopping.LSAtT — Type

Type: LSAtT

Methods: update!, reinit!, copy

A structure designed to track line search information from one iteration to another. Given f : ℜⁿ → ℜ, define h(θ) = f(x + θ*d) where x and d are vectors of same dimension and θ is a scalar, more specifically the step size.

Tracked data can include:

x : the current step size

ht [opt] : h(θ) at the current iteration

gt [opt] : h'(θ)

h₀ [opt] : h(0)

g₀ [opt] : h'(0)

current_time [opt] : the time at which the line search algorithm started.

current_score [opt] : the score at which the line search algorithm started.

Constructor: LSAtT(:: Number; ht :: FloatVoid = nothing, gt :: FloatVoid = nothing, h₀ :: FloatVoid = nothing, g₀ :: FloatVoid = nothing, current_time :: FloatVoid = nothing, current_score :: Iterate = nothing)

Note: By default, unknown entries are set to nothing.

General Functions

Stopping.update! — Function

update!: generic update function for the State

update!(:: AbstractState; convert = false, kwargs...)

The function compares the kwargs and the entries of the State. If the type of the kwargs is the same as the entry or the entry is nothing, then it is updated.

Set kargs convert to true to update even incompatible types.

Examples: update!(state1) update!(state1, current_time = 2.0) update!(state1, convert = true, current_time = 2.0)

See also: GenericState, reinit!, update_and_start!, update_and_stop!

Stopping.reinit! — Function

reinit!: function that set all the entries at nothing except the mandatory x.

reinit!(:: AbstractState, :: Iterate; kwargs...)

Note: If x is given as a kargs it will be prioritized over the second argument.

Examples: reinit!(state2, zeros(2)) reinit!(state2, zeros(2), current_time = 1.0)

There is a shorter version of reinit! reusing the x in the state

reinit!(:: AbstractState; kwargs...)

Examples: reinit!(state2) reinit!(state2, current_time = 1.0)

reinit!: function that set all the entries at void except the mandatory x

reinit!(:: NLPAtX, x :: AbstractVector, l :: AbstractVector; kwargs...)

reinit!(:: NLPAtX; kwargs...)

Note: if x, lambda or evals are given as keyword arguments they will be prioritized over the existing x, lambda and the default Counters.

reinit!: reinitialize the MetaData in the Stopping.

reinit!(:: AbstractStopping; rstate :: Bool = false, kwargs...)

Note:

If rstate is set as true it reinitializes the current State

(with the kwargs).

If rlist is set as true the list of states is also reinitialized, either

set as nothing if rstate is true, and a list containing only the current state if rstate is false.

Missing docstring.

Missing docstring for Stopping.copy, Stopping.compress_state!, Stopping.copy_compress_state. Check Documenter's build log for details.

Stopping.add_to_list! — Function

add_to_list!: add a State to the list of maximal size n. If a n+1-th State is added, the first one in the list is removed. The given is State is compressed before being added in the list (via State.copy_compress_state).

add_to_list!(:: ListStates, :: AbstractState; kwargs...)

Note: kwargs are passed to the compress_state call.

see also: ListStates, State.compress_state, State.copy_compress_state

Base.length — Function

length: return the number of States in the list.

length(:: ListStates)

see also: print, addtolist!, ListStates

Base.print — Function

print: output formatting. return a DataFrame.

print(:: ListStates; verbose :: Bool = true, print_sym :: Union{Nothing,Array{Symbol,1}})

Note:

set verbose to false to avoid printing.

if print_sym is an Array of Symbol, only those symbols are printed. Note that

the returned DataFrame still contains all the columns.

More information about DataFrame: http://juliadata.github.io/DataFrames.jl

see also: add_to_list!, length, ListStates

Stopping

Types

Stopping.GenericStopping — Type

Type: GenericStopping

Methods: start!, stop!, update_and_start!, update_and_stop!, fill_in!, reinit!, status

A generic Stopping to solve instances with respect to some optimality conditions. Optimality is decided by computing a score, which is then tested to zero.

Tracked data include:

pb : A problem

state : The information relative to the problem, see GenericState

(opt) meta : Metadata relative to a stopping criterion, see StoppingMeta.

(opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

(opt) listofstates : ListStates designed to store the history of States.

(opt) userspecificstruct : Contains any structure designed by the user.

Constructor: GenericStopping(:: Any, :: AbstractState; meta :: AbstractStoppingMeta = StoppingMeta(), main_stp :: Union{AbstractStopping, Nothing} = nothing, user_specific_struct :: Any = nothing, kwargs...)

Note: Metadata can be provided by the user or created with the Stopping constructor via kwargs. If a specific StoppingMeta is given and kwargs are provided, the kwargs have priority.

Examples: GenericStopping(pb, GenericState(ones(2)), rtol = 1e-1)

Besides optimality conditions, we consider classical emergency exit: - domain error (for instance: NaN in x) - unbounded problem (not implemented) - unbounded x (x is too large) - tired problem (time limit attained) - resources exhausted (not implemented) - stalled problem (not implemented) - iteration limit (maximum number of iteration (i.e. nb of stop) attained) - main_pb limit (tired or resources of main problem exhausted)

There is an additional default constructor which creates a Stopping with a default State.

GenericStopping(:: Any, :: Union{Number, AbstractVector}; kwargs...)

Note: Keywords arguments are forwarded to the classical constructor.

Examples: GenericStopping(pb, x0, rtol = 1e-1)

Stopping.NLPStopping — Type

Type: NLPStopping

Methods: start!, stop!, update_and_start!, update_and_stop!, fill_in!, reinit!, status KKT, unconstrained_check, unconstrained2nd_check, optim_check_bounded

Specialization of GenericStopping. Stopping structure for non-linear programming problems using NLPModels.

Attributes:

pb : an AbstractNLPModel

state : The information relative to the problem, see GenericState

(opt) meta : Metadata relative to stopping criterion, see StoppingMeta.

(opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

(opt) listofstates : ListStates designed to store the history of States.

(opt) userspecificstruct : Contains any structure designed by the user.

NLPStopping(:: AbstractNLPModel, :: AbstractState; meta :: AbstractStoppingMeta = StoppingMeta(), max_cntrs :: Dict = _init_max_counters(), main_stp :: Union{AbstractStopping, Nothing} = nothing, list :: Union{ListStates, Nothing} = nothing, user_specific_struct :: Any = nothing, kwargs...)

Note:

designed for NLPAtX State. Constructor checks that the State has the

required entries.

There is an additional default constructor creating a Stopping where the State is by default and the optimality function is the function KKT().

NLPStopping(pb :: AbstractNLPModel; kwargs...)

Note: Kwargs are forwarded to the classical constructor.

Stopping.LS_Stopping — Type

Type: LS_Stopping (specialization of GenericStopping)

Methods: start!, stop!, update_and_start!, update_and_stop!, fill_in!, reinit!, status, armijo, wolfe, armijo_wolfe, shamanskii_stop, goldstein

Specialization of GenericStopping. LS_Stopping is designed to handle the stopping criterion of line search problems. Let f:R→Rⁿ, then h(t) = f(x+td) where x and d are vectors and t is a scalar. h is such that h:R→R.

Stopping structure for 1D non-linear programming problems. Input :

pb : an Any

state : The information relative to the problem, see GenericState

(opt) meta : Metadata relative to stopping criterion.

(opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

(opt) listofstates : ListStates designed to store the history of States.

(opt) userspecificstruct : Contains any structure designed by the user.

LS_Stopping(:: Any, :: LSAtT; meta :: AbstractStoppingMeta = StoppingMeta(), main_stp :: Union{AbstractStopping, Nothing} = nothing, user_specific_struct :: Any = nothing, kwargs...)

Note:

The pb can be a LineModel defined in SolverTools.jl (https://github.com/JuliaSmoothOptimizers/SolverTools.jl)

It is possible to define those stopping criterion in a NLPStopping except NLPStopping uses vectors operations. LS_Stopping and it's admissible functions (Armijo and Wolfe are provided with Stopping.jl) uses scalar operations.

optimality_check(pb, state; kwargs...) -> Float64 is by default armijo For instance, the armijo condition is: h(t)-h(0)-τ₀th'(0) ⩽ 0 therefore armijo(h, hatt) returns the maximum between h(t)-h(0)-τ₀th'(0) and 0.

See also GenericStopping, NLPStopping, LSAtT

Stopping.LAStopping — Type

Type: LAStopping

Methods: start!, stop!, update_and_start!, update_and_stop!, fillin!, reinit!, status linear\system_check, normal_equation_check

Specialization of GenericStopping. Stopping structure for linear algebra solving either

$Ax = b$

or

$min\_{x} \tfrac{1}{2}\|Ax - b\|^2$.

Attributes:

pb : a problem using LLSModel (designed for linear least square problem, see https://github.com/JuliaSmoothOptimizers/NLPModels.jl/blob/master/src/lls_model.jl )

state : The information relative to the problem, see GenericState

(opt) meta : Metadata relative to stopping criterion, see StoppingMeta.

(opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

(opt) listofstates : ListStates designed to store the history of States.

(opt) userspecificstruct : Contains any structure designed by the user.

LAStopping(:: LLSModel, :: AbstractState; meta :: AbstractStoppingMeta = StoppingMeta() main_stp :: Union{AbstractStopping, Nothing} = nothing, user_specific_struct :: Any = nothing, kwargs...)

Note:

Kwargs are forwarded to the classical constructor.

Not specific State targeted

State don't necessarily keep track of evals

Evals are checked only for pb.A being a LinearOperator

zero_start is true if 0 is the initial guess (not check automatically)

LLSModel counter follow NLSCounters (see initmaxcountersNLS in NLPStoppingmod.jl)

By default, meta.max_cntrs is initialized with an NLSCounters

There is additional constructors:

LAStopping(:: Union{AbstractLinearOperator, AbstractMatrix}, :: AbstractVector, kwargs...) LAStopping(:: Union{AbstractLinearOperator, AbstractMatrix}, :: AbstractVector, :: AbstractState, kwargs...)

See also GenericStopping, NLPStopping, LS_Stopping, linear_system_check, normal_equation_check

Stopping.LACounters — Type

Type: LACounters

Stopping.StoppingMeta — Type

Type: StoppingMeta

Methods: no methods.

Attributes:

atol : absolute tolerance.

rtol : relative tolerance.

optimality0 : optimality score at the initial guess.

tol_check : Function of atol, rtol and optimality0 testing a score to zero.

optimality_check : a stopping criterion via an admissibility function

unbounded_threshold : threshold for unboundedness of the problem.

unbounded_x : threshold for unboundedness of the iterate.

max_f : maximum number of function (and derivatives) evaluations.

max_cntrs : Dict contains the maximum number of evaluations

max_eval : maximum number of function (and derivatives) evaluations.

max_iter : threshold on the number of stop! call/number of iteration.

max_time : time limit to let the algorithm run.

nb_of_stop : keep track of the number of stop! call/iteration.

start_time : keep track of the time at the beginning.

fail_sub_pb : status.

unbounded : status.

unbounded_pb : status.

tired : status.

stalled : status.

iteration_limit : status.

resources : status.

optimal : status.

infeasible : status.

main_pb : status.

domainerror : status.

suboptimal : status.

StoppingMeta(;atol :: Number = 1.0e-6, rtol :: Number = 1.0e-15, optimality0 :: Number = 1.0, tol_check :: Function = (atol,rtol,opt0) -> max(atol,rtol*opt0), unbounded_threshold :: Number = 1.0e50, unbounded_x :: Number = 1.0e50, max_f :: Int = typemax(Int), max_eval :: Int = 20000, max_iter :: Int = 5000, max_time :: Number = 300.0, start_time :: Float64 = NaN, kwargs...)

Note:

It is a mutable struct, therefore we can modify elements of a StoppingMeta.

The nb_of_stop is incremented everytime stop! or update_and_stop! is called

The optimality0 is modified once at the beginning of the algorithm (start!)

The start_time is modified once at the beginning of the algorithm (start!) if not precised before.

The different status: fail_sub_pb, unbounded, unbounded_pb, tired, stalled, iteration_limit, resources, optimal, main_pb, domainerror, suboptimal, infeasible

fail_sub_pb, suboptimal, and infeasible are modified by the algorithm.

optimality_check takes two inputs (AbstractNLPModel, NLPAtX)

and returns a Number to be compared to 0.

optimality_check does not necessarily fill in the State.

Examples: StoppingMeta()

General Functions

Stopping.start! — Function

start!: update the Stopping and return a boolean true if we must stop.

start!(:: AbstractStopping; no_start_opt_check :: Bool = false, kwargs...)

Purpose is to know if there is a need to even perform an optimization algorithm or if we are at an optimal solution from the beginning. Set no_start_opt_check to true to avoid checking optimality.

The function start! successively calls: _domain_check, _optimality_check, _null_test

Note: - start! initialize the start_time (if not done before) and meta.optimality0. - Keywords argument are sent to the _optimality_check! call.

Stopping.update_and_start! — Function

update_and_start!: update the values in the State and initialize the Stopping. Returns the optimality status of the problem as a boolean.

update_and_start!(:: AbstractStopping; kwargs...)

Note: Kwargs are forwarded to the update! call.

Stopping.stop! — Function

stop!: update the Stopping and return a boolean true if we must stop.

stop!(:: AbstractStopping; kwargs...)

It serves the same purpose as start! in an algorithm; telling us if we stop the algorithm (because we have reached optimality or we loop infinitely, etc).

The function stop! successively calls: _domain_check, _optimality_check, _null_test, _unbounded_check!, _tired_check!, _resources_check!, _stalled_check!, _iteration_check!, _main_pb_check!, add_to_list!

Note:

Kwargs are sent to the _optimality_check! call.

If listofstates != nothing, call add_to_list! to update the list of State.

Stopping.update_and_stop! — Function

update_and_stop!: update the values in the State and return the optimality status of the problem as a boolean.

update_and_stop!(stp :: AbstractStopping; kwargs...)

Note: Kwargs are forwarded to the update! call.

Missing docstring.

Missing docstring for Stopping.reinit!. Check Documenter's build log for details.

Stopping.fill_in! — Function

fill_in!: fill in the unspecified values of the AbstractState.

fill_in!(:: AbstractStopping, x :: Union{Number, AbstractVector})

Note: NotImplemented for Abstract/Generic-Stopping.

fill_in!: (NLPStopping version) a function that fill in the required values in the NLPAtX

fill_in!( :: NLPStopping, :: Iterate; fx :: Iterate = nothing, gx :: Iterate = nothing, Hx :: Iterate = nothing, cx :: Iterate = nothing, Jx :: Iterate = nothing, lambda :: Iterate = nothing, mu :: Iterate = nothing, matrix_info :: Bool = true, kwargs...)

Stopping.status — Function

status: returns the status of the algorithm:

status(:: AbstractStopping; list = false)

The different status are:

Optimal: reached an optimal solution.

Unbounded: current iterate too large in norm.

UnboundedPb: unbouned problem.

Stalled: stalled algorithm.

IterationLimit: too many iterations of the algorithm.

Tired: algorithm too slow.

ResourcesExhausted: too many ressources used, i.e. too many functions evaluations.

ResourcesOfMainProblemExhausted: in the case of a substopping, ResourcesExhausted or Tired for the main stopping.

Infeasible: default return value, if nothing is done the problem is considered feasible.

DomainError: there is a NaN somewhere.

Note:

Set keyword argument list to true, to get an Array with all the status.

The different status correspond to boolean values in the MetaData, see StoppingMeta.

Non-linear admissibility functions

Stopping.KKT — Function

KKT: verifies the KKT conditions

KKT( :: AbstractNLPModel, :: NLPAtX; pnorm :: Float64 = Inf, kwargs...)

Note: state.gx is mandatory + if bounds state.mu + if constraints state.cx, state.Jx, state.lambda.

See also unconstrained_check, unconstrained2nd_check, optimcheckbounded

Stopping.unconstrained_check — Function

unconstrained: return the infinite norm of the gradient of the objective function

unconstrained_check( :: AbstractNLPModel, :: NLPAtX; pnorm :: Float64 = Inf, kwargs...)

required: state.gx (filled if nothing)

See also unconstrained2nd_check, optimcheckbounded, KKT

Stopping.unconstrained2nd_check — Function

unconstrained 2nd: check the norm of the gradient and the smallest eigenvalue of the hessian.

unconstrained2nd_check( :: AbstractNLPModel, :: NLPAtX; pnorm :: Float64 = Inf, kwargs...)

Note: required are state.gx, state.Hx (filled if nothing).

See also unconstrained_check, optimcheckbounded, KKT

Stopping.optim_check_bounded — Function

optim_check_bounded: gradient of the objective function projected

optim_check_bounded( :: AbstractNLPModel, :: NLPAtX; pnorm :: Float64 = Inf, kwargs...)

Note: required is state.gx (filled if nothing)

See also unconstrained_check, unconstrained2nd_check, KKT

Linear algebra admissibility functions

Stopping.linear_system_check — Function

linear_system_check: return ||Ax-b||_p

linear_system_check(:: Any, :: AbstractState; pnorm :: Float64 = Inf, kwargs...)

Note:

Returns the p-norm of state.res

state.res is filled in if nothing.

Stopping.normal_equation_check — Function

linear_system_check: return ||A'Ax-A'b||_p

linear_system_check(:: Any, :: AbstractState; pnorm :: Float64 = Inf, kwargs...)

Note: pb must have A and b entries

Line search admissibility functions

Stopping.armijo — Function

armijo: check if a step size is admissible according to the Armijo criterion.

Armijo criterion: f(x + θd) - f(x) - τ₀ θ ∇f(x+θd)d < 0

armijo(h :: Any, h_at_t :: LSAtT; τ₀ :: Float64 = 0.01, kwargs...)

Note: ht, h₀ and g₀ are required in the LSAtT

See also wolfe, armijo_wolfe, shamanskii_stop, goldstein

Stopping.wolfe — Function

wolfe: check if a step size is admissible according to the Wolfe criterion.

Strong Wolfe criterion: |∇f(x+θd)| < τ₁||∇f(x)||.

wolfe(h :: Any, h_at_t :: LSAtT; τ₁ :: Float64 = 0.99, kwargs...)

Note: gt and g₀ are required in the LSAtT

See also armijo, armijo_wolfe, shamanskii_stop, goldstein

Stopping.armijo_wolfe — Function

armijo_wolfe: check if a step size is admissible according to the Armijo and Wolfe criteria.

armijo_wolfe(h :: Any, h_at_t :: LSAtT; τ₀ :: Float64 = 0.01, τ₁ :: Float64 = 0.99, kwargs...)

Note: ht, h₀, gt and g₀ are required in the LSAtT

See also armijo, wolfe, shamanskii_stop, goldstein

Stopping.shamanskii_stop — Function

shamanskii_stop: check if a step size is admissible according to the "Shamanskii" criteria. This criteria was proposed in: Lampariello, F., & Sciandrone, M. (2001). Global convergence technique for the Newton method with periodic Hessian evaluation. Journal of optimization theory and applications, 111(2), 341-358.

shamanskii_stop(h :: Any, h_at_t :: LSAtT; γ :: Float64 = 1.0e-09, kwargs...)

Note: * h.d accessible (specific LineModel) * ht, h₀ are required in the LSAtT

See also armijo, wolfe, armijo_wolfe, goldstein

Stopping.goldstein — Function

goldstein: check if a step size is admissible according to the Goldstein criteria.

goldstein(h :: Any, h_at_t :: LSAtT; τ₀ :: Float64 = 0.0001, τ₁ :: Float64 = 0.9999, kwargs...)

Note: ht, h₀ and g₀ are required in the LSAtT

See also armijo, wolfe, armijo_wolfe, shamanskii_stop