State

Types

Missing docstring.

Missing docstring for Stopping.GenericStatemod. Check Documenter's build log for details.

Stopping.NLPAtX — Type

Type: NLPAtX Methods: update!, reinit!

NLPAtX contains the information concerning a nonlinear problem at the iteration x. min_{x ∈ ℜⁿ} f(x) subject to lcon <= c(x) <= ucon, lvar <= x <= uvar.

Basic information is:

x : the current candidate for solution to our original problem

fx : which is the funciton evaluation at x

gx : which is the gradient evaluation at x

Hx : which is the hessian representation at x

mu : Lagrange multiplier of the bounds constraints

cx : evaluation of the constraint function at x

Jx : jacobian matrix of the constraint function at x

lambda : Lagrange multiplier of the constraints

current_time : time

current_score : score

evals : number of evaluations of the function (import the type NLPModels.Counters)

Note: * by default, unknown entries are set to nothing (except evals). * All these information (except for x and lambda) are optionnal and need to be update when required. The update is done trhough the update! function. * x and lambda are mandatory entries. If no constraints lambda = []. * The constructor check the size of the entries.

Stopping.LSAtT — Type

Type: LSAtT Methods: update!, reinit!, copy

A structure designed to track line search information from one iteration to another. If we have f : ℜⁿ → ℜ, then we define h(θ) = f(x + θ*d) where x and d are vectors of same dimension and θ is a scalar, more specifically our step size.

Tracked data can include:

x : our current step size

ht : h(θ) at the current iteration

gt : h'(θ)

h₀ : h(0)

g₀ : h'(0)

current_time: the time at which the line search algorithm started.

current_score: the score at which the line search algorithm started.

Note: by default, unknown entries are set to nothing.

General Functions

Stopping.update! — Function

update!: generic update function for the State

The function compare the kwargs and the entries of the State. If the type of the kwargs is the same as the entry or the entry was nothing, then it is updated.

Set convert to true, to update even incompatible types.

Stopping.reinit! — Function

reinit!: function that set all the entries at void except the mandatory x

Note: If x is given as a keyword argument it will be prioritized over the argument x.

reinit!: shorter version of reinit! reusing the x in the state

Note: If x is given as a keyword argument it will be prioritized over the argument x.

reinit!: function that set all the entries at void except the mandatory x

Warning: if x, lambda or evals are given as a keyword argument they will be prioritized over the existing x, lambda and the default Counters.

reinit!: short version of reinit! reusing the x in the state

Warning: if x, lambda or evals are given as a keyword argument they will be prioritized over the existing x, lambda and the default Counters.

reinit!: Reinitialize the meta data.

If rstate is set as true it reinitialize the current_state (with the kwargs)

Stopping

Types

Stopping.GenericStopping — Type

Type: GenericStopping Methods: start!, stop!, updateandstart!, updateandstop!, fill_in!, reinit!, status

A generic stopping criterion to solve instances (pb) with respect to some optimality conditions. Optimality is decided by computing a score, which is then tested to zero.

Besides optimality conditions, we consider classical emergency exit:

domain error (for instance: NaN in x)

unbounded problem (not implemented)

unbounded x (x is too large)

tired problem (time limit attained)

resources exhausted (not implemented)

stalled problem (not implemented)

iteration limit (maximum number of iteration (i.e. nb of stop) attained)

main_pb limit (tired or resources of main problem exhausted)

Input : - pb : A problem - state : The information relative to the problem, see GenericState - (opt) meta : Metadata relative to stopping criterion. - (opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

Note: Metadata can be provided by the user or created with the Stopping constructor with kwargs. If a specific StoppingMeta is given and kwargs are provided, the kwargs have priority.

Stopping.NLPStopping — Type

Type: NLPStopping (specialization of GenericStopping) Methods: start!, stop!, updateandstart!, updateandstop!, fill_in!, reinit!, status

Stopping structure for non-linear programming problems using NLPModels. Input : - pb : an AbstractNLPModel - optimalitycheck : a stopping criterion through an admissibility function - state : The information relative to the problem, see GenericState - maxcntrs : Dict contains the max number of evaluations - (opt) meta : Metadata relative to stopping criterion. - (opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

Note:

optimality_check : takes two inputs (AbstractNLPModel, NLPAtX)

and returns a Float64 to be compared at 0.

designed for NLPAtX State. Constructor checks that the State has the

required entries.

Warning:

optimality_check does not necessarily fill in the State.

Stopping.LS_Stopping — Type

Type: LSStopping (specialization of GenericStopping) Methods: start!, stop!, updateandstart!, updateandstop!, fillin!, reinit!, status

LS_Stopping is designed to handle the stopping criterion of line search problems. Let f:R→Rⁿ, then h(t) = f(x+td) where x and d are vectors and t is a scalar. h is such that h:R→R.

Stopping structure for non-linear programming problems using NLPModels. Input : - pb : an AbstractNLPModel - optimalitycheck : a stopping criterion through an admissibility function - state : The information relative to the problem, see GenericState - (opt) meta : Metadata relative to stopping criterion. - (opt) mainstp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then nothing.

Note:

The pb can be a LineModel defined in SolverTools.jl (https://github.com/JuliaSmoothOptimizers/SolverTools.jl)

It is possible to define those stopping criterion in a NLPStopping except NLPStopping uses vectors operations. LS_Stopping and it's admissible functions (Armijo and Wolfe are provided with Stopping.jl) uses scalar operations.

optimalitycheck(pb, state; kwargs...) -> Float64 For instance, the armijo condition is: h(t)-h(0)-τ₀th'(0) ⩽ 0 therefore armijo(h, hat_t) returns the maximum between h(t)-h(0)-τ₀th'(0) and 0.

Stopping.StoppingMeta — Type

Type: StoppingMeta Methods: no

Common stopping criteria for algorithms: - absolute and relative tolerance - threshold for unboundedness - time limit to let the algorithm run - maximum number of function (and derivatives) evaluations

It's a mutable struct therefore we can modified elements of a StoppingMeta. - The nbofstop is incremented everytime stop! or updateandstop! is called - The optimality0 is modified once at the beginning of the algorithm (start!) - The starttime is modified once at the beginning of the algorithm (start!) if not precised before. - The different status: failsubpb, unbounded, unboundedpb, tired, stalled, iterationlimit, resources, optimal, mainpb, domainerror, suboptimal, infeasible

Note: failsubpb, suboptimal, and infeasible are modified by the algorithm.

General Functions

Stopping.start! — Function

start!: update the Stopping and return a boolean true if we must stop.

Purpose is to know if there is a need to even perform an optimization algorithm or if we are at an optimal solution from the beginning.

Note: * start! initialize the starttime (if not done before) and meta.optimality0. * Keywords argument are sent to the _optimalitycheck!

Stopping.update_and_start! — Function

updateandstart!: Update the values in the State and initializes the Stopping Returns the optimity status of the problem.

Kwargs are forwarded to the update!(stp.current_state)

Stopping.stop! — Function

stop!: update the Stopping and return a boolean true if we must stop.

serves the same purpose as start! in an algorithm, tells us if we stop the algorithm (because we have reached optimality or we loop infinitely, etc).

Keywords argument are sent to the optimalitycheck!

Stopping.update_and_stop! — Function

updateandstop!: update the values in the State and returns the optimality status of the problem.

kwargs are forwarded to the update!(stp.current_state)

Missing docstring.

Missing docstring for Stopping.reinit!. Check Documenter's build log for details.

Stopping.fill_in! — Function

fill_in!: A function that fill in the unspecified values of the AbstractState.

NotImplemented for Abstract/Generic-Stopping.

fill_in!: a function that fill in the required values in the State

Stopping.status — Function

status: returns the status of the algorithm: - Optimal : if we reached an optimal solution - Unbounded : if the problem doesn't have a lower bound - Stalled : if algorithm is stalling - IterationLimit : if we did too many iterations of the algorithm - Tired : if the algorithm takes too long - ResourcesExhausted: if we used too many ressources, i.e. too many functions evaluations - ResourcesOfMainProblemExhausted: in the case of a substopping, ResourcesExhausted or Tired for the main stopping. - Infeasible : default return value, if nothing is done the problem is considered infeasible - DomainError : there is a NaN somewhere

Note: set keyword argument list to true, to get an Array with all the status.

Non linear admissibility functions

Stopping.KKT — Function

KKT: verifies the KKT conditions

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

Stopping.unconstrained_check — Function

unconstrained: return the infinite norm of the gradient of the objective function

required: state.gx (filled if nothing)

Stopping.unconstrained2nd_check — Function

unconstrained 2nd: check the norm of the gradient and the smallest eigenvalue of the hessian.

required: state.gx, state.Hx (filled if nothing)

Stopping.optim_check_bounded — Function

optimcheckbounded: gradient of the objective function projected

required: state.gx (filled if void)

Line search admissibility functions

Stopping.armijo — Function

armijo: check if a step size is admissible according to the Armijo criterion.

Armijo criterion: f(x + θd) - f(x) < τ₀ θ ∇f(x+θd)d

Note: ht, h₀ and g₀ are required in the LSAtT

Stopping.wolfe — Function

wolfe: check if a step size is admissible according to the Wolfe criterion.

Strong Wolfe criterion: |∇f(x+θd)| < τ₁||∇f(x)||.

Note: gt and g₀ are required in the LSAtT

Stopping.armijo_wolfe — Function

armijo_wolfe: check if a step size is admissible according to the Armijo and Wolfe criteria.

Note: ht, h₀, gt and g₀ are required in the LSAtT

Stopping.shamanskii_stop — Function

shamanskii_stop: check if a step size is admissible according to the "Shamanskii" criteria. This criteria was proposed in GLOBAL CONVERGENCE TECHNIQUE FOR THE NEWTON METHOD WITH PERIODIC HESSIAN EVALUATION by F. LAMPARIELLO and M. SCIANDRONE

Note: * h.d accessible (specific LineModel) * ht, h₀ are required in the LSAtT

Stopping.goldstein — Function

goldstein: check if a step size is admissible according to the Goldstein criteria.

Note: ht, h₀ and g₀ are required in the LSAtT