Types

Stopping.GenericStopping — Type

Type : GenericStopping Methods : start!, stop!

A generic stopping criterion to solve instances (pb) with respect to some optimality conditions (optimality_check) Besides optimality conditions, we consider classical emergency exit:

unbounded problem

stalled problem

tired problem (measured by the number of evaluations of functions and time)

Input : - pb : An problem - state : The information relative to the problem - (opt) meta : Metadata relative to stopping criterion. Can be provided by the user or created with the Stopping constructor with kwargs If a specific StoppingMeta is given as well as kwargs are provided, the kwargs have priority. - (opt) main_stp : Stopping of the main loop in case we consider a Stopping of a subproblem. If not a subproblem, then of type Nothing.

Stopping.NLPStopping — Type

Stopping structure for non-linear programming problems. Inputs:

pb : An AbstractNLPModel

optimality_check : a stopping criterion through an admissibility function

meta : StoppingMeta

max_cntrs :: Dict contains the max number of evaluations

current_state : the current state of the problem (i.e an NLPAtX)

optimality_check : takes two inputs (AbstractNLPModel, NLPAtX)

and returns a Float64 to be compared at 0. (Idée: ajouter une nouvelle input mainpb dans le optimalitycheck?)

Stopping.LS_Stopping — Type

LSStopping is designed to handle the stopping criterion of line search problems. Let f:R→Rⁿ, then h(t) = f(x+td) where x and d are vectors and t is a scalar. h is such that h:R→R. h is a LineModel defined in SolverTools.jl (https://github.com/JuliaSmoothOptimizers/SolverTools.jl) It is possible to define those stopping criterion in a NLPStopping except NLPStopping uses vectors operations. LSStopping and it's admissible functions (Armijo and Wolfe are provided with Stopping.jl) uses scalar operations. In order to work properly within the Stopping framework, admissible functions must return a value that will be compare to 0. For instance, the armijo condition is h(t)-h(0)-τ₀th'(0) ⩽ 0 therefore armijo(h, hatt) returns the maximum between h(t)-h(0)-τ₀th'(0) and 0. The inputs of an admissible function are : - h :: A LineModel - hatt :: A line search state, defined in State.jl

Stopping.StoppingMeta — Type

StoppingMeta Common stopping criterion for "all" optimization algorithms such as: - absolute and relative tolerance - threshold for unboundedness - time limit to let the algorithm run - maximum number of function (and derivatives) evaluations

It's a mutable struct therefore we can modified elements of a StoppingMeta. - The nbofstop is incremented everytime stop! or updateandstop! is called - The optimality0 is modified once at the beginning of the algorithm (start!) - The starttime is modified once at the beginning of the algorithm (start!) if not precised before. - The different status failsub_pb, unbounded, tired, stalled, optimal, suboptimal, and infeasible are modified according to the data of the algorithm.

General Functions

Stopping.update_and_start! — Function

updateandstart!: Update the values in the State and initializes the Stopping Returns the optimity status of the problem.

Stopping.update_and_stop! — Function

updateandstop!: Update the values in the State and returns the optimity status of the problem.

Stopping.fill_in! — Function

fill_in!: A function that fill in the unspecified values of the AbstractState.

fill_in! : A function that fill in the required values in the State

Stopping.start! — Function

start!: Input: Stopping. Output: optimal or not. Purpose is to know if there is a need to even perform an optimization algorithm or if we are at an optimal solution from the beginning.

Note: start! initialize the start_time (if not done before) and meta.optimality0.

Keywords argument are sent to the optimalitycheck!

Stopping.stop! — Function

stop!: Inputs: Interface Stopping. Output: optimal or not. Serves the same purpose as start! When in an algorithm, tells us if we stop the algorithm (because we have reached optimality or we loop infinitely, etc).

Keywords argument are sent to the optimalitycheck!

Stopping.reinit! — Function

reinit!: Input: Stopping. Output: Stopping modified. Reinitialize the meta data filled in by the start!

Stopping.status — Function

status: Takes an AbstractStopping as input. Returns the status of the algorithm: - Optimal : if we reached an optimal solution - Unbounded : if the problem doesn't have a lower bound - Stalled : if we did too many iterations of the algorithm - Tired : if the algorithm takes too long - ResourcesExhausted: if we used too many ressources, i.e. too many functions evaluations - ResourcesOfMainProblemExhausted: in the case of a substopping, ResourcesExhausted or Tired for the main stopping. - Infeasible : default return value, if nothing is done the problem is considered infeasible - DomainError : there is a NaN somewhere

Non linear admissibility functions

Stopping.unconstrained — Function

unconstrained: return the infinite norm of the gradient of the objective function

required: state.gx (filled if void)

Stopping.KKT — Function

KKT: verifies the KKT conditions

required: state.gx

if bounds: state.mu

if constraints: state.cx, state.Jx, state.lambda

Line search admissibility functions

Stopping.armijo — Function

Check if a step size is admissible according to the Armijo criterion. Inputs: Any, #LineModel and LSAtT. Outputs: admissibility in the armijo sense Armijo criterion: f(x + θd) - f(x) < τ₀ ∇f(x+θd)d

Stopping.wolfe — Function

Check if a step size is admissible according to the Wolfe criterion. Inputs: Any, #LineModel and LSAtT. Outputs: admissibility in the Strong Wolfe sense. Strong Wolfe criterion: |∇f(x+θd)| < τ₁||∇f(x)||.

Stopping.armijo_wolfe — Function

Check if a step size is admissible according to both the Armijo and Wolfe criterion. Inputs: Any, #LineModel and LSAtT. Outputs: admissibility in the Strong Wolfe sense. Strong Wolfe criterion: |∇f(x+θd)| < τ₁||∇f(x)||.